requirements.txt
telethon>=1.25.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0
